FIX: addresses PR change requests
