+++ app/components/MeatballMenu/MenuItems/index.ts
@@ -1,0 +1,1 @@
+export { MenuItems } from './MenuItems';
